+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import { ClerkProvider } from "@clerk/nextjs";
+import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
 import "./globals.css";
 import { TRPCProvider } from "@/trpc/client";
@@ -26,6 +27,9 @@ export default function RootLayout({
           <TRPCProvider>
             <ToasterProvider />
             {children}
+            {process.env.NODE_ENV === "development" && (
+              <ReactQueryDevtools initialIsOpen={false} />
+            )}
           </TRPCProvider>
         </body>
       </html>